Move SQLCipher support into separate file
